+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Add additional output when storing artifacts (#207)
 - Fix longer (2+) chains of runs that have source-dir specified (#151)
+- Output more descriptive error message when setup environment fails (#230)
 
 ## v1.0.560 (2016-07-14)